requirements.txt
flask
python-dotenv
google-generativeai
gtts
pillow
requests
flask-cors
